+++ tests/PuliTemplateLoaderTest.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\Extension\Twig\Tests;
+namespace Puli\TwigExtension\Tests;
 
 use PHPUnit_Framework_MockObject_MockObject;
 use PHPUnit_Framework_TestCase;
-use Puli\Extension\Twig\PuliTemplateLoader;
 use Puli\Repository\Api\ResourceRepository;
 use Puli\Repository\Tests\Resource\TestDirectory;
+use Puli\TwigExtension\PuliTemplateLoader;
 
 /**
  * @since  1.0